Cleanup project files
