+++ Assets/Models/Ships.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 8b49336e087b93f489b2764130f7bdb4
-NativeFormatImporter:
+guid: 7dca7708e2eaec7688f2a72f65ff9fe7
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 15600000
   userData: 
   assetBundleName: 
   assetBundleVariant:
