+++ Assets/Models/Ships/00_SandboxShip.fbx.meta
