+++ Assets/Prefabs/Input Manager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1607c2235b60063d09cd5ea87becde89
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
