+++ Assets/Scripts/Generated.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 50d931a9d514dbbc99aaa0c1ab247a88
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
