+++ ProjectSettings/TagManager.asset
@@ -15,8 +15,8 @@ TagManager:
   - 
   - 
   - 
-  - Player 1
-  - Player 2
+  - 
+  - 
   - 
   - 
   -
